# gamification-impl/src/main/resources/api-spec.yaml
openapi: '3.0.3'
info:
  version: '0.1.0'
  title: Gamification API
  description: description
security:
  - apiKeyAuth: []
paths:
...
  /badges:
    post:
      description: create badge for the website
      operationId: createBadge
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Badge'
      responses:
        '201':
          description: created
          content:
            'application/json':
              schema:
                type: string
        '400':
          description: parameters are empty or null
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '409':
          description: badge already exists
    get:
      description: get the list of all badges
      operationId: getBadges
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Badge'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
...
  /users:
    get:
      description: get the list of all users
      operationId: getUsers
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: user not found
  /users/{userAppId}:
    get:
      description: get one user, by userAppId
      operationId: getUser
      parameters:
        - in: path
          name: userAppId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    #    bearerAuth:            # arbitrary name for the security scheme
    #      type: http
    #      scheme: bearer       # warning: it HAS TO be a lowercase b for the swagger ui version used in springfox
    #      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
...
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
  schemas:
...
    Badge:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    User:
      type: object
      properties:
        userAppId:
          type: string
        points:
          type: integer
        badges:
          type: array
          items:
            $ref: '#/components/schemas/Badge'
    UserInfo:
      type: object
      properties:
        points:
          type: integer
        badges:
          type: array
          items:
            $ref: '#/components/schemas/Badge'
...